Split agent runner boundaries
