+++ Sources/QuillCodeAgent/AgentRawTextStreamActionCollector.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension AgentRunner {
+    static func collectStreamingAction(from stream: AsyncThrowingStream<String, Error>) async throws -> AgentAction {
+        try await AgentActionStreamCollector.collect(
+            from: stream,
+            emptyError: AgentError.emptyStreamingResponse
+        )
+    }
+}
