+++ Sources/QuillCodeAgent/AgentTypes.swift
@@ -1,0 +1,69 @@
+import Foundation
+import QuillCodeCore
+
+public enum AgentAction: Sendable, Hashable {
+    case say(String)
+    case tool(ToolCall)
+}
+
+public protocol LLMClient: Sendable {
+    func nextAction(thread: ChatThread, userMessage: String, tools: [ToolDefinition]) async throws -> AgentAction
+}
+
+public protocol StreamingLLMClient: LLMClient {
+    func actionTextStream(
+        thread: ChatThread,
+        userMessage: String,
+        tools: [ToolDefinition]
+    ) async throws -> AsyncThrowingStream<String, Error>
+}
+
+public protocol UsageStreamingLLMClient: StreamingLLMClient {
+    func actionEventStream(
+        thread: ChatThread,
+        userMessage: String,
+        tools: [ToolDefinition]
+    ) async throws -> AsyncThrowingStream<AgentTextStreamEvent, Error>
+}
+
+public enum AgentError: Error, CustomStringConvertible {
+    case emptyStreamingResponse
+    case promisedWorkWithoutToolAction
+    case tooManyToolSteps(Int)
+
+    public var description: String {
+        switch self {
+        case .emptyStreamingResponse:
+            return "The model stream finished without returning an action."
+        case .promisedWorkWithoutToolAction:
+            return "The model promised to perform work but did not return a tool action."
+        case .tooManyToolSteps(let limit):
+            return "The agent reached the tool-step limit (\(limit)) before returning a final answer."
+        }
+    }
+}
+
+public struct AgentRunResult: Sendable {
+    public var thread: ChatThread
+    public var toolResults: [ToolResult]
+
+    public init(thread: ChatThread, toolResults: [ToolResult]) {
+        self.thread = thread
+        self.toolResults = toolResults
+    }
+}
+
+public struct AgentToolFeedback: Codable, Sendable, Hashable {
+    public var toolCall: ToolCall
+    public var result: ToolResult
+    public var followUpResult: ToolResult?
+
+    public init(toolCall: ToolCall, result: ToolResult, followUpResult: ToolResult? = nil) {
+        self.toolCall = toolCall
+        self.result = result
+        self.followUpResult = followUpResult
+    }
+}
+
+public typealias AgentRunProgressHandler = @Sendable (ChatThread) async -> Void
+public typealias AgentToolExecutionOverride = @Sendable (ToolCall, URL) async -> ToolResult?
