+++ Tests/QuillCodeParityTests/ParityAgentGateTests.swift
@@ -60,16 +60,34 @@ final class ParityAgentGateTests: QuillCodeParityTestCase {
 
     func testAgentStreamingHelpersLiveOutsideAgentRunnerFile() throws {
         let agentText = try Self.agentSourceText(named: "Agent.swift")
+        let actionResolverText = try Self.agentSourceText(named: "AgentActionResolver.swift")
         let streamingText = try Self.agentSourceText(named: "AgentActionStreaming.swift")
+        let textRunnerText = try Self.agentSourceText(named: "AgentTextStreamActionRunner.swift")
+        let usageRunnerText = try Self.agentSourceText(named: "AgentUsageStreamActionRunner.swift")
+        let rawCollectorText = try Self.agentSourceText(named: "AgentRawTextStreamActionCollector.swift")
+        let textCollectorText = try Self.agentSourceText(named: "AgentTextStreamActionCollector.swift")
+        let usageCollectorText = try Self.agentSourceText(named: "AgentUsageStreamActionCollector.swift")
+        let draftPublisherText = try Self.agentSourceText(named: "AgentStreamingDraftPublisher.swift")
 
         XCTAssertTrue(streamingText.contains("public enum AgentActionStreamCollector"), "Streaming action collection should live in a focused helper.")
         XCTAssertTrue(streamingText.contains("public enum AgentActionStreamPreview"), "Partial assistant preview parsing should live with streaming helpers.")
         XCTAssertTrue(streamingText.contains("var rawActionText"), "Progressive stream accumulation should live with the stream collector.")
         XCTAssertTrue(streamingText.contains("AgentActionStreamPreview.visibleAssistantText"), "Stream collector should own draft-preview extraction.")
-        XCTAssertTrue(agentText.contains("AgentActionStreamCollector.collect"), "AgentRunner should delegate streaming collection.")
+        XCTAssertTrue(actionResolverText.contains("nextTextStreamingAction"), "Agent action resolution should delegate text-streaming setup.")
+        XCTAssertTrue(actionResolverText.contains("nextUsageStreamingAction"), "Agent action resolution should delegate usage-streaming setup.")
+        XCTAssertTrue(textRunnerText.contains("func nextTextStreamingAction"), "Text streaming setup should live outside Agent.swift.")
+        XCTAssertTrue(usageRunnerText.contains("func nextUsageStreamingAction"), "Usage streaming setup should live outside Agent.swift.")
+        XCTAssertTrue(rawCollectorText.contains("AgentActionStreamCollector.collect"), "Raw streaming collection should delegate to the focused collector.")
+        XCTAssertTrue(textCollectorText.contains("AgentActionStreamCollector.collect"), "Draft-updating text streams should delegate to the focused collector.")
+        XCTAssertTrue(usageCollectorText.contains("AgentActionStreamCollector.collect"), "Usage streams should delegate to the focused collector.")
+        XCTAssertTrue(draftPublisherText.contains("publishAssistantDraft"), "Streaming draft publication should live outside Agent.swift.")
+        XCTAssertTrue(draftPublisherText.contains("publishReasoningSummary"), "Reasoning summary publication should live outside Agent.swift.")
         XCTAssertFalse(agentText.contains("public enum AgentActionStreamCollector"), "Agent.swift should not own streaming collection details.")
         XCTAssertFalse(agentText.contains("private static func partialJSONStringValue"), "Agent.swift should not own partial JSON preview parsing.")
         XCTAssertFalse(agentText.contains("AgentActionStreamPreview.visibleAssistantText"), "Agent.swift should not own streaming preview parsing.")
+        XCTAssertFalse(agentText.contains("AgentActionStreamCollector.collect"), "Agent.swift should not directly collect streaming actions.")
+        XCTAssertFalse(agentText.contains("actionTextStream("), "Agent.swift should not own text-streaming setup.")
+        XCTAssertFalse(agentText.contains("actionEventStream("), "Agent.swift should not own usage-streaming setup.")
         XCTAssertFalse(agentText.contains("var rawActionText"), "Agent.swift should not own raw streaming accumulation.")
     }
